Many additions -- SortCup Object++
* updated types involving resistances to doubles or longs.
* Added SortCup object which handles logic involving cups
* completed main loop logic (some changes may be necessary for
additional features)
* added feedEmpty function on the StepFeed.

+++ src/RS_Mainboard/ProgmemData.cpp
@@ -26,7 +26,7 @@
 // Constants for Calibrated Measurements (Voltages, Resistances, Currents)
 	const double avHigh = 3.3;
 	const double avLow = 1.5;
-	const int internalTestResistances[6] = {10000000, 1000000, 100000, 10000, 1000, 100};
+	const long internalTestResistances[6] = {10000000, 1000000, 100000, 10000, 1000, 100};
 	const double internalCurrentSources[3] = {0.1, 0.02937, 0.01818};
 
 // Standard Resistor values
